[refactor] 모달 버튼 파일 분리 #7
