+++ frontend/lib/widgets/user_details.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:frontend/widgets/profile_img.dart';
+import 'package:frontend/widgets/modal_button.dart';
 import 'package:frontend/widgets/choose_purpose.dart';
 
 class UserDetails extends StatelessWidget {
@@ -93,43 +94,6 @@ class UserDetails extends StatelessWidget {
                 },
               ),
             ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class ModalButton extends StatelessWidget {
-  final String text;
-  final Function handlePressed;
-
-  const ModalButton({
-    super.key,
-    required this.text,
-    required this.handlePressed,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 280,
-      height: 50,
-      child: ElevatedButton(
-        onPressed: () {
-          handlePressed();
-        },
-        style: ElevatedButton.styleFrom(
-          backgroundColor: const Color(0xff212121),
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(10),
-          ),
-        ),
-        child: Text(
-          text,
-          style: const TextStyle(
-            color: Colors.white,
-            fontSize: 18,
           ),
         ),
       ),
